requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2019.11.28
chardet==3.0.4
discord.py==1.6.0
idna==2.7
multidict==4.7.5
mysql==0.0.2
mysqlclient==2.0.3
neotermcolor==2.0.8
pycodestyle==2.6.0
pymongo==3.11.2
requests==2.25.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==6.0
yarl==1.4.2
